feat(installation): изменена структура проекта

+++ quantitizer/__init__.py
@@ -1,10 +1,9 @@
 import numpy as np
 from scipy.cluster.vq import kmeans2
 
-from pq_array import PQ
+from quantitizer._pq_array import PQ
 
 
-# http://ethen8181.github.io/machine-learning/deep_learning/multi_label/product_quantization.html
 def quantitize(vectors, sub_size=8, n_cluster=256, n_iter=20, minit='points', seed=123):
     if len(vectors[0]) % sub_size != 0:
         raise Exception(f"sub_size должен нацело делить {len(vectors[0])}")
@@ -19,8 +18,8 @@ def quantitize(vectors, sub_size=8, n_cluster=256, n_iter=20, minit='points', se
         code_books.append(centroid)
         indexes.append(label)
 
-    code_books = np.array(code_books)
-    indexes = np.array(indexes).T
+    code_books = np.array(code_books, dtype=np.float16)
+    indexes = np.array(indexes, dtype=np.int8).T
 
     # return code_books, np.array(indexes).T
     return PQ(
@@ -28,8 +27,6 @@ def quantitize(vectors, sub_size=8, n_cluster=256, n_iter=20, minit='points', se
         n_cluster, indexes, code_books)
 
 
-# TODO: не работает
-# https://github.com/avidale/compress-fasttext/blob/master/compress_fasttext/quantization.py
 def quantitize_experimental(vectors, sub_size=8, sample=None, n_cluster=256, n_iter=20, minit='points', seed=123):
     encoder = pqkmeans.encoder.PQEncoder(
         iteration=n_iter,
